+++ include/circt/Scheduling/SimplexSchedulers.h
@@ -1,0 +1,301 @@
+//===- SimplexSchedulers.h - Linear programming-based schedulers --------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef CIRCT_SIMPLEX_SCHEDULERS_H
+#define CIRCT_SIMPLEX_SCHEDULERS_H
+
+#include "circt/Scheduling/Utilities.h"
+
+#include "mlir/IR/Operation.h"
+
+namespace circt::scheduling {
+
+/// This class provides a framework to model certain scheduling problems as
+/// lexico-parametric linear programs (LP), which are then solved with an
+/// extended version of the dual simplex algorithm.
+///
+/// The approach is described in:
+///  [1] B. D. de Dinechin, "Simplex Scheduling: More than Lifetime-Sensitive
+///      Instruction Scheduling", PRISM 1994.22, 1994.
+///  [2] B. D. de Dinechin, "Fast Modulo Scheduling Under the Simplex Scheduling
+///      Framework", PRISM 1995.01, 1995.
+///
+/// Resource-free scheduling problems (called "central problems" in the papers)
+/// have an *integer* linear programming formulation with a totally unimodular
+/// constraint matrix. Such ILPs can however be solved optimally in polynomial
+/// time with a (non-integer) LP solver (such as the simplex algorithm), as the
+/// LP solution is guaranteed to be integer. Note that this is the same idea as
+/// used by SDC-based schedulers.
+class SimplexSchedulerBase {
+protected:
+  /// The objective is to minimize the start time of this operation.
+  Operation *lastOp;
+
+  /// S is part of a mechanism to assign fixed values to the LP variables.
+  int parameterS;
+
+  /// T represents the initiation interval (II). Its minimally-feasible value is
+  /// computed by the algorithm.
+  int parameterT;
+
+  /// The simplex tableau is the algorithm's main data structure.
+  /// The dashed parts always contain the zero respectively the identity matrix,
+  /// and therefore are not stored explicitly.
+  ///
+  ///                        ◀───nColumns────▶
+  ///           nParameters────┐
+  ///                        ◀─┴─▶
+  ///                       ┌─────┬───────────┬ ─ ─ ─ ─ ┐
+  ///                      ▲│. . .│. . ... . .│    0        ▲
+  ///           nObjectives││. . .│. . ... . .│         │   │
+  ///                      ▼│. . .│. . ... . .│             │
+  ///                       ├─────┼───────────┼ ─ ─ ─ ─ ┤   │
+  ///  firstConstraintRow > │. . .│. . ... . .│1            │nRows
+  ///                       │. . .│. . ... . .│  1      │   │
+  ///                       │. . .│. . ... . .│    1        │
+  ///                       │. . .│. . ... . .│      1  │   │
+  ///                       │. . .│. . ... . .│        1    ▼
+  ///                       └─────┴───────────┴ ─ ─ ─ ─ ┘
+  ///       parameter1Column ^
+  ///         parameterSColumn ^
+  ///           parameterTColumn ^
+  ///  firstNonBasicVariableColumn ^
+  ///                              ─────────── ──────────
+  ///                       nonBasicVariables   basicVariables
+  SmallVector<SmallVector<int>> tableau;
+
+  /// During the pivot operation, one column in the elided part of the tableau
+  /// is modified; this vector temporarily catches the changes.
+  SmallVector<int> implicitBasicVariableColumnVector;
+
+  /// The linear program models the operations' start times as variables, which
+  /// we identify here as 0, ..., |ops|-1.
+  /// Additionally, for each dependence (precisely, the inequality modeling the
+  /// precedence constraint), a slack variable is required; these are identified
+  /// as |ops|, ..., |ops|+|deps|-1.
+  ///
+  /// This vector stores the numeric IDs of non-basic variables. A variable's
+  /// index *i* in this vector corresponds to the tableau *column*
+  /// `firstNonBasicVariableColumn`+*i*.
+  SmallVector<unsigned> nonBasicVariables;
+
+  /// This vector store the numeric IDs of basic variables. A variable's index
+  /// *i* in this vector corresponds to the tableau *row*
+  /// `firstConstraintRow`+*i*.
+  SmallVector<unsigned> basicVariables;
+
+  /// Used to conveniently retrieve an operation's start time variable. The
+  /// alternative would be to find the op's index in the problem's list of
+  /// operations.
+  DenseMap<Operation *, unsigned> startTimeVariables;
+
+  /// This vector keeps track of the current locations (i.e. row or column) of
+  /// a start time variable in the tableau. We encode column numbers as positive
+  /// integers, and row numbers as negative integers. We do not track the slack
+  /// variables.
+  SmallVector<int> startTimeLocations;
+
+  /// Non-basic variables can be "frozen" to a specific value, which prevents
+  /// them from being pivoted into basis again.
+  DenseMap<unsigned, unsigned> frozenVariables;
+
+  /// Number of rows in the tableau = |obj| + |deps|.
+  unsigned nRows;
+  /// Number of explicitly stored columns in the tableau = |params| + |ops|.
+  unsigned nColumns;
+
+  // Number of objective rows.
+  unsigned nObjectives;
+  /// All other rows encode linear constraints.
+  unsigned &firstConstraintRow = nObjectives;
+
+  // Number of parameters (fixed for now).
+  static constexpr unsigned nParameters = 3;
+  /// The first column corresponds to the always-one "parameter" in u = (1,S,T).
+  static constexpr unsigned parameter1Column = 0;
+  /// The second column corresponds to the variable-freezing parameter S.
+  static constexpr unsigned parameterSColumn = 1;
+  /// The third column corresponds to the parameter T, i.e. the current II.
+  static constexpr unsigned parameterTColumn = 2;
+  /// All other (explicitly stored) columns represent non-basic variables.
+  static constexpr unsigned firstNonBasicVariableColumn = nParameters;
+
+  /// Allow subclasses to collect additional constraints that are not part of
+  /// the input problem, but should be modeled in the linear problem.
+  SmallVector<Problem::Dependence> additionalConstraints;
+
+  virtual Problem &getProblem() = 0;
+  virtual LogicalResult checkLastOp();
+  virtual bool fillObjectiveRow(SmallVector<int> &row, unsigned obj);
+  virtual void fillConstraintRow(SmallVector<int> &row,
+                                 Problem::Dependence dep);
+  virtual void fillAdditionalConstraintRow(SmallVector<int> &row,
+                                           Problem::Dependence dep);
+  void buildTableau();
+
+  int getParametricConstant(unsigned row);
+  SmallVector<int> getObjectiveVector(unsigned column);
+  std::optional<unsigned> findDualPivotRow();
+  std::optional<unsigned> findDualPivotColumn(unsigned pivotRow,
+                                              bool allowPositive = false);
+  std::optional<unsigned> findPrimalPivotColumn();
+  std::optional<unsigned> findPrimalPivotRow(unsigned pivotColumn);
+  void multiplyRow(unsigned row, int factor);
+  void addMultipleOfRow(unsigned sourceRow, int factor, unsigned targetRow);
+  void pivot(unsigned pivotRow, unsigned pivotColumn);
+  LogicalResult solveTableau();
+  LogicalResult restoreDualFeasibility();
+  bool isInBasis(unsigned startTimeVariable);
+  unsigned freeze(unsigned startTimeVariable, unsigned timeStep);
+  void translate(unsigned column, int factor1, int factorS, int factorT);
+  LogicalResult scheduleAt(unsigned startTimeVariable, unsigned timeStep);
+  void moveBy(unsigned startTimeVariable, unsigned amount);
+  unsigned getStartTime(unsigned startTimeVariable);
+
+  void dumpTableau();
+
+public:
+  explicit SimplexSchedulerBase(Operation *lastOp) : lastOp(lastOp) {}
+  virtual ~SimplexSchedulerBase() = default;
+  virtual LogicalResult schedule() = 0;
+};
+
+/// This class solves the basic, acyclic `Problem`.
+class SimplexScheduler : public SimplexSchedulerBase {
+private:
+  Problem &prob;
+
+protected:
+  Problem &getProblem() override { return prob; }
+
+public:
+  SimplexScheduler(Problem &prob, Operation *lastOp)
+      : SimplexSchedulerBase(lastOp), prob(prob) {}
+
+  LogicalResult schedule() override;
+};
+
+/// This class solves the resource-free `CyclicProblem`.  The optimal initiation
+/// interval (II) is determined as a side product of solving the parametric
+/// problem, and corresponds to the "RecMII" (= recurrence-constrained minimum
+/// II) usually considered as one component in the lower II bound used by modulo
+/// schedulers.
+class CyclicSimplexScheduler : public SimplexSchedulerBase {
+private:
+  CyclicProblem &prob;
+
+protected:
+  Problem &getProblem() override { return prob; }
+  void fillConstraintRow(SmallVector<int> &row,
+                         Problem::Dependence dep) override;
+
+public:
+  CyclicSimplexScheduler(CyclicProblem &prob, Operation *lastOp)
+      : SimplexSchedulerBase(lastOp), prob(prob) {}
+  LogicalResult schedule() override;
+};
+
+// This class solves acyclic, resource-constrained `SharedOperatorsProblem` with
+// a simplified version of the iterative heuristic presented in [2].
+class SharedOperatorsSimplexScheduler : public SimplexSchedulerBase {
+private:
+  SharedOperatorsProblem &prob;
+
+protected:
+  Problem &getProblem() override { return prob; }
+
+public:
+  SharedOperatorsSimplexScheduler(SharedOperatorsProblem &prob,
+                                  Operation *lastOp)
+      : SimplexSchedulerBase(lastOp), prob(prob) {}
+  LogicalResult schedule() override;
+};
+
+// This class solves the `ModuloProblem` using the iterative heuristic presented
+// in [2].
+class ModuloSimplexScheduler : public CyclicSimplexScheduler {
+private:
+  struct MRT {
+    ModuloSimplexScheduler &sched;
+
+    using TableType = SmallDenseMap<unsigned, DenseSet<Operation *>>;
+    using ReverseTableType = SmallDenseMap<Operation *, unsigned>;
+    SmallDenseMap<Problem::OperatorType, TableType> tables;
+    SmallDenseMap<Problem::OperatorType, ReverseTableType> reverseTables;
+
+    explicit MRT(ModuloSimplexScheduler &sched) : sched(sched) {}
+    LogicalResult enter(Operation *op, unsigned timeStep);
+    void release(Operation *op);
+  };
+
+  ModuloProblem &prob;
+  SmallVector<unsigned> asapTimes, alapTimes;
+  SmallVector<Operation *> unscheduled, scheduled;
+  MRT mrt;
+
+protected:
+  Problem &getProblem() override { return prob; }
+  LogicalResult checkLastOp() override;
+  enum { OBJ_LATENCY = 0, OBJ_AXAP /* i.e. either ASAP or ALAP */ };
+  bool fillObjectiveRow(SmallVector<int> &row, unsigned obj) override;
+  void updateMargins();
+  void scheduleOperation(Operation *n);
+  unsigned computeResMinII();
+
+public:
+  ModuloSimplexScheduler(ModuloProblem &prob, Operation *lastOp)
+      : CyclicSimplexScheduler(prob, lastOp), prob(prob), mrt(*this) {}
+  LogicalResult schedule() override;
+};
+
+// This class solves the `ChainingProblem` by relying on pre-computed
+// chain-breaking constraints.
+class ChainingSimplexScheduler : public SimplexSchedulerBase {
+private:
+  ChainingProblem &prob;
+  float cycleTime;
+
+protected:
+  Problem &getProblem() override { return prob; }
+  void fillAdditionalConstraintRow(SmallVector<int> &row,
+                                   Problem::Dependence dep) override;
+
+public:
+  ChainingSimplexScheduler(ChainingProblem &prob, Operation *lastOp,
+                           float cycleTime)
+      : SimplexSchedulerBase(lastOp), prob(prob), cycleTime(cycleTime) {}
+  LogicalResult schedule() override;
+};
+
+// This class solves the resource-free `ChainingCyclicProblem` by relying on
+// pre-computed chain-breaking constraints. The optimal initiation interval (II)
+// is determined as a side product of solving the parametric problem, and
+// corresponds to the "RecMII" (= recurrence-constrained minimum II) usually
+// considered as one component in the lower II bound used by modulo schedulers.
+class ChainingCyclicSimplexScheduler : public SimplexSchedulerBase {
+private:
+  ChainingCyclicProblem &prob;
+  float cycleTime;
+
+protected:
+  Problem &getProblem() override { return prob; }
+  void fillConstraintRow(SmallVector<int> &row,
+                         Problem::Dependence dep) override;
+  void fillAdditionalConstraintRow(SmallVector<int> &row,
+                                   Problem::Dependence dep) override;
+
+public:
+  ChainingCyclicSimplexScheduler(ChainingCyclicProblem &prob, Operation *lastOp,
+                                 float cycleTime)
+      : SimplexSchedulerBase(lastOp), prob(prob), cycleTime(cycleTime) {}
+  LogicalResult schedule() override;
+};
+
+}; // namespace circt::scheduling
+
+#endif
